+++ src/client/js/nologin.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'unstated';
 import { I18nextProvider } from 'react-i18next';
 
-import i18nFactory from './util/i18n';
+import { i18nFactory } from './util/i18n';
 
 import AppContainer from './services/AppContainer';